+++ changelog.md
@@ -4,8 +4,7 @@ Changelog
 0.4.0 (????-??-??)
 ------------------
 
-* Implemented the 'format' command.
-* Implemented "parse" command.
+* Implemented the "format", "parse" and "release" commands.
 
 
 0.3.0 (2023-02-12)